backend/requirements.txt
fastapi
uvicorn
pyht
python-dotenv
openai-whisper
